Merge SipProcessingFailure into SipProcessingException.
A review of the Fairfax ingestion documentation reveals that different
processing issues are treated as exceptions rather than failures (since
some states are not failures, but just not routine.

The merging of SipProcessingFailure into SipProcessingException.groovy means
that the code is consistent with the ingestion design.

This has involved some refactoring of multiple classes.

- Renamed SipProcessingFailureReason to SipProcessingExceptionReason.
- Renamed SipProcessingExceptionReasonType to SipProcessingExceptionType.
- Added 'fatal' attribute to SipProcessingExceptionReasonType.
- Moved SipProcessingException into subpackage state.